+++ docs/CHANGELOG.md
@@ -3,6 +3,16 @@ Small (C) Black Rook Software 2020
 by Matt Tropiano et al. (see AUTHORS.txt)
 
 
+Changed in [NOW]
+----------------
+
+- `Added` SmallResponse.
+- `Added` Lots of additional methods to SmallResponseUtils.
+- `Added` SmallUtils.sendContent().
+- `Added` Controllers do not need View/Attachment/Content annotations for methods if the return type is SmallResponse.
+- `Changed` Unified the content send.
+
+
 Changed in 1.0.0
 ----------------